// prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TournamentStatus = {
  REGISTRATION: 'REGISTRATION',
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type TournamentStatus = (typeof TournamentStatus)[keyof typeof TournamentStatus]


export const PlayerStatus = {
  ONLINE: 'ONLINE',
  OFFLINE: 'OFFLINE',
  INGAME: 'INGAME',
  AWAY: 'AWAY'
} as const

export type PlayerStatus = (typeof PlayerStatus)[keyof typeof PlayerStatus]


export const PlayerTrend = {
  UP: 'UP',
  DOWN: 'DOWN',
  STABLE: 'STABLE'
} as const

export type PlayerTrend = (typeof PlayerTrend)[keyof typeof PlayerTrend]


export const AchievementRarity = {
  COMMON: 'COMMON',
  RARE: 'RARE',
  EPIC: 'EPIC',
  LEGENDARY: 'LEGENDARY'
} as const

export type AchievementRarity = (typeof AchievementRarity)[keyof typeof AchievementRarity]